+++ app/zoosper-page/config/controllers.php
@@ -173,4 +173,3 @@ return [
         adminUrls: $services->get(AdminUrlGenerator::class),
     ),
 ];
-